requirements.txt
torch==2.1.2
torchvision
torchsummary
tqdm
PyYAML
gdown==4.6.0
Flask
Flask-Cors
mlflow==2.9.2
dvc
joblib
notebook
